+++ src/test/java/com/dsi/parallax/pipeline/text/TestStringSequenceToNGramsPipe.java
@@ -32,7 +32,7 @@ public class TestStringSequenceToNGramsPipe {
             assertTrue(it.hasNext());
             List<String> vec = it.next().getData();
             assertTrue(!it.hasNext());
-            assertEquals(vec.size(),(int)Math.max(0, tokens2.size()-(i-1)));
+            assertEquals(vec.size(),Math.max(0, tokens2.size()-(i-1)));
         }
         
         for(int i = 1; i < 105; i++) {
@@ -44,7 +44,7 @@ public class TestStringSequenceToNGramsPipe {
             assertTrue(it.hasNext());
             List<String> vec = it.next().getData();
             assertTrue(!it.hasNext());
-            assertEquals(vec.size(),(int)Math.max(0, longTokens.size()-(i-1)));
+            assertEquals(vec.size(),Math.max(0, longTokens.size()-(i-1)));
         }
     }
